+++ firmware/CHANGELOG.md
@@ -37,6 +37,7 @@ Release template (copy/paste this for new release):
  - Ford Voodoo 1-5-4-8-3-7-2-6 firing order
  - 6/0 CKP trigger pattern for 1995 Lamborghini Diablo #5876
  - VE/Ign/Boost blend tables can select a Y axis override for even more flexibility
+ - Disable radiator fans after certain vehicle speed #5875
 
 ### Removed
  - Narrow to Wideband approximation